+++ CMakeLists.txt
@@ -258,6 +258,7 @@ set(TESTS
         db/db_test.cc
         db/db_compaction_filter_test.cc
         db/db_dynamic_level_test.cc
+        db/db_universal_compaction_test.cc
         db/db_tailing_iter_test.cc
         db/dbformat_test.cc
         db/deletefile_test.cc
